The set_color() function has been renamed to set_static_color().
Now, when entering a command, not `PLUSTERM/main` is displayed on the left, but just PT.

There is a new function set_color(), it does not work yet (in development).

The new "web" command (opens the site).

The "pt" command has been deleted (temporarily).

The version in the code has been replaced with a new one (0.20.1).